fixes en get review

+++ src/com/lyu/inoodles/logic/Review.java
@@ -6,36 +6,36 @@ import com.lyu.inoodles.infrastructure.ReviewData;
 
 public class Review {
 
-    int mFlav;
-    int mSpic;
-    int mOverall;
+    float mFlav;
+    float mSpic;
+    float mOverall;
     List<String> mComments;
 
     public static Review getReviewByFkNoodles(int fkNoodles) {
         return ReviewData.getReviewByFkNoodles(fkNoodles);
     }
 
-    public int getFlav() {
+    public float getFlav() {
         return mFlav;
     }
 
-    public void setFlav(int mFlav) {
+    public void setFlav(float mFlav) {
         this.mFlav = mFlav;
     }
 
-    public int getSpic() {
+    public float getSpic() {
         return mSpic;
     }
 
-    public void setSpic(int mSpic) {
+    public void setSpic(float mSpic) {
         this.mSpic = mSpic;
     }
 
-    public int getOverall() {
+    public float getOverall() {
         return mOverall;
     }
 
-    public void setOverall(int mOverall) {
+    public void setOverall(float mOverall) {
         this.mOverall = mOverall;
     }
